Move ConfigureLearningStandards(IWindsorContainer from StartupBase to CommonConfigurationInstaller so that it can also be visible to .NET Core Startup.
